perf(photography): ~140x lighter page — optimized images, manifest dims, ISR
Before: client fetch + spinner, then 169 original JPEGs (6–21 MB each;
~191 MB transferred for the initial viewport) with per-image relayout.

- Serve tiles/lightbox via next/image (Vercel image optimization + srcset);
  initial viewport now ~1.4 MB
- Precompute image dimensions into gallery-meta/manifest.json
  (scripts/build-gallery-manifest.mjs, header-range fetch + sharp) so the
  justified grid is stable from first paint; on-load measuring kept only as
  a batched fallback for unmanifested photos
- /photography is now a server component with revalidate=300 passing data
  to the client grid (no fetch waterfall, no spinner)
- /api/gallery moved to shared @mtosity/lib/gallery + CDN cache headers
- Debounced resize relayout
