+++ packages/lib/package.json
@@ -6,6 +6,7 @@
   "exports": {
     "./constants": "./src/constants.ts",
     "./db": "./src/db.ts",
+    "./gallery": "./src/gallery.ts",
     "./notes": "./src/notes.ts",
     "./rate-limit": "./src/rate-limit.ts",
     "./jsonld": "./src/jsonld.ts"
@@ -16,6 +17,7 @@
   "dependencies": {
     "@neondatabase/serverless": "^1.1.0",
     "@upstash/ratelimit": "^2.0.8",
+    "@vercel/blob": "^2.0.0",
     "@upstash/redis": "^1.37.0"
   },
   "peerDependencies": {
